user entity has been added

+++ internal/repository/task_repository.go
@@ -9,10 +9,10 @@ import (
 
 type TaskRepository interface {
 	Create(task *domain.Task) error
-	GetById(id int64) (*domain.Task, error)
-	GetAll() ([]domain.Task, error)
-	Update(*domain.Task) error
-	Delete(id int64) error
+	GetById(id int64, userID int64) (*domain.Task, error)
+	GetAll(userID int64) ([]domain.Task, error)
+	Update(task *domain.Task, userID int64) error
+	Delete(id int64, userID int64) error
 }
 
 type taskRepository struct {
@@ -24,33 +24,33 @@ func NewRepository(db *sqlx.DB) TaskRepository {
 }
 
 func (t *taskRepository) Create(task *domain.Task) error {
-	row := t.db.QueryRowx("INSERT INTO tasks(title, description) VALUES($1, $2) RETURNING id, created_at", task.Title, task.Description)
+	row := t.db.QueryRowx("INSERT INTO tasks(title, description, user_id) VALUES($1, $2, $3) RETURNING id, created_at", task.Title, task.Description, task.UserId)
 	if err := row.Scan(&task.ID, &task.CreatedAt); err != nil {
 		return err
 	}
 	return nil
 }
 
-func (t *taskRepository) GetById(id int64) (*domain.Task, error) {
+func (t *taskRepository) GetById(id int64, userID int64) (*domain.Task, error) {
 	var task domain.Task
-	err := t.db.Get(&task, "SELECT id, title, description, created_at, completed FROM tasks WHERE id=$1", id)
+	err := t.db.Get(&task, "SELECT id, title, description, created_at, completed FROM tasks WHERE id=$1 AND user_id=$2", id, userID)
 	if err != nil {
 		return nil, err
 	}
 	return &task, nil
 }
 
-func (t *taskRepository) GetAll() ([]domain.Task, error) {
+func (t *taskRepository) GetAll(userID int64) ([]domain.Task, error) {
 	var tasks []domain.Task
-	err := t.db.Select(&tasks, "SELECT id, title, description, created_at, completed FROM tasks")
+	err := t.db.Select(&tasks, "SELECT id, title, description, created_at, completed FROM tasks WHERE user_id=$1", userID)
 	if err != nil {
 		return nil, err
 	}
 	return tasks, nil
 }
 
-func (t *taskRepository) Update(task *domain.Task) error {
-	result, err := t.db.Exec("UPDATE tasks set title=$1, description=$2, completed=$3 WHERE id = $4", task.Title, task.Description, task.Completed, task.ID)
+func (t *taskRepository) Update(task *domain.Task, userID int64) error {
+	result, err := t.db.Exec("UPDATE tasks set title=$1, description=$2, completed=$3 WHERE id = $4 AND user_id=$5", task.Title, task.Description, task.Completed, task.ID, userID)
 	if err != nil {
 		return err
 	}
@@ -66,8 +66,8 @@ func (t *taskRepository) Update(task *domain.Task) error {
 	return nil
 }
 
-func (t *taskRepository) Delete(id int64) error {
-	result, err := t.db.Exec("DELETE FROM tasks WHERE id=$1", id)
+func (t *taskRepository) Delete(id int64, userID int64) error {
+	result, err := t.db.Exec("DELETE FROM tasks WHERE id=$1 AND user_id=$2", id, userID)
 	if err != nil {
 		return err
 	}
